requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.5
fonttools==4.55.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pillow==11.0.0
playwright==1.49.1
playwright-stealth==1.0.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
pyee==12.0.0
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rfc3986==2.0.0
rich==13.7.1
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
tomli==2.0.1
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
